Fix signed vs unsigned comparison.

+++ src/hook-recvfrom.c
@@ -27,8 +27,8 @@ static FilterReplyResult filter_parse_reply(FilterReplyResultBase * const rb,
         if (obj_nbyte != NULL &&
             obj_nbyte->type == MSGPACK_OBJECT_POSITIVE_INTEGER) {
             const int64_t new_nbyte = obj_nbyte->via.i64;
-            if (new_nbyte <= INT_MAX && new_nbyte <= *nbyte) {
-                *nbyte = new_nbyte;
+            if (new_nbyte <= INT_MAX && (size_t) new_nbyte <= *nbyte) {
+                *nbyte = (size_t) new_nbyte;
             }
         }
         const msgpack_object * const obj_flags =
